requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-postgres>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0